docked-screenshot-taker/requirements.txt
fastmcp
yt-dlp
python-dotenv 